requirements.txt
django-colorfield
django-flat-theme
django-flat-responsive
six>=1.9.0
